requirements.txt
git+git://github.com/openclimatefix/nowcasting_dataset@main#egg=nowcasting_dataset
numpy
pandas
matplotlib
zarr
xarray
ipykernel
h5netcdf
gcsfs
scikit-image
torch
pytorch-lightning
perceiver_pytorch
dask
pvlib
pyproj
pytest
pytest-cov
flake8
jedi
mypy
tables
tilemapbase
neptune-client
boto3
rich
omegaconf
hydra-core
python-dotenv
hydra-optuna-sweeper
psutil
plotly